refactor: comprehensive code quality improvements across core package
This commit includes multiple code quality optimizations and refactoring:

## Code Organization & Modularity

### Extract isIgnoredFile to shared utilities (Optimization 6)
- Moved isIgnoredFile function from transform/index.ts to shared/utils.ts
- Added comprehensive JSDoc documentation
- Improved code reusability across the project
- Updated type definitions

### Extract Pug processing logic (Optimization 7)
- Created new file: transform/transform-vue-pug.ts (187 lines)
- Separated Pug-specific logic from main Vue transform
- Reduced transform-vue.ts from 221 to ~100 lines
- Added helper functions: isPugTemplate, calculateLineOffsets, transformPugTemplate
- Improved code maintainability and testability

## Code Quality Improvements

### Replace @ts-ignore with @ts-expect-error (Optimization 10)
- Replaced 9 instances of @ts-ignore with @ts-expect-error
- Added explanatory comments for each type suppression
- Improved type safety and maintainability
- Files affected: transform-jsx.ts (4), transform-vue-pug.ts (5)

### Optimize event listeners management (Optimization 11)
- Centralized event listener configuration in array
- Added attachEventListeners() and detachEventListeners() helper methods
- Reduced code duplication from ~39 lines to ~46 lines (but more maintainable)
- Improved code organization and extensibility

### Extract magic numbers to constants (Optimization 12)
- Added POPPER_MARGIN constant (value: 10)
- Replaced hardcoded values with named constant
- Improved code readability and maintainability

### Optimize line counting algorithm (Optimization 13)
- Refactored countLines function in transform-svelte.ts
- Changed from O(n) loop to built-in split method
- Reduced from 9 lines to 1 line implementation
- Added TypeScript type annotation and JSDoc

### Add comprehensive JSDoc documentation (Optimization 14)
- Added detailed JSDoc for 4 complex functions in shared/utils.ts:
  * isDev: Environment detection with priority logic
  * matchCondition: Flexible file matching with multiple condition types
  * getMappingFilePath: Path mapping resolution with examples
  * isExcludedFile: File exclusion logic with include/exclude patterns
- Improved IDE intellisense and developer experience

## Bug Fixes

### Fix missing semicolon (Optimization 8)
- Fixed missing semicolon at client/index.ts:464
- Maintained code style consistency

## Test Improvements

### Enhance showNotification test coverage
- Added 3 comprehensive test cases for requestAnimationFrame behavior
- Improved coverage for line 632 (classList.add animation trigger)
- Tests now verify:
  * requestAnimationFrame is called correctly
  * Show class is added after animation frame
  * Multiple notifications handle separate animation frames
- Increased test count from 12 to 15 tests

## Statistics

- Files changed: 10
- Lines added: 565
- Lines removed: 254
- New files: 2 (transform-vue-pug.ts, transform-vue-pug.d.ts)
- Test coverage maintained: 99.96%
- All 852 tests passing
