+++ packages/core/src/shared/utils.ts
@@ -95,7 +95,35 @@ export function getDependencies() {
 }
 
 type BooleanFunction = () => boolean;
-// 判断当前是否为 development 环境; 优先判定用户指定的环境；其次判断系统默认的环境
+
+/**
+ * Determine if the current environment is development mode
+ *
+ * Priority: user-specified environment > system default environment
+ *
+ * @param userDev - User-specified development mode setting:
+ *   - `true`: Force development mode
+ *   - `false`: Force production mode
+ *   - `function`: Dynamic function that returns boolean
+ *   - `undefined`: Use system default
+ * @param systemDev - System default development mode (e.g., from NODE_ENV)
+ * @returns `true` if in development mode, `false` otherwise
+ *
+ * @example
+ * ```typescript
+ * // Force development mode
+ * isDev(true, false) // Returns: true
+ *
+ * // Force production mode
+ * isDev(false, true) // Returns: false
+ *
+ * // Use system default
+ * isDev(undefined, true) // Returns: true
+ *
+ * // Dynamic function
+ * isDev(() => process.env.NODE_ENV === 'development', false)
+ * ```
+ */
 export function isDev(
   userDev: boolean | BooleanFunction | undefined,
   systemDev: boolean
@@ -112,6 +140,33 @@ export function isDev(
   return dev || systemDev;
 }
 
+/**
+ * Check if a file matches the given condition
+ *
+ * Supports multiple condition types for flexible file matching:
+ * - String: Checks if file path contains the string
+ * - RegExp: Tests file path against the regular expression
+ * - Array: Recursively checks if file matches any condition in the array
+ *
+ * @param condition - The condition to match against:
+ *   - `string`: File path must contain this string
+ *   - `RegExp`: File path must match this pattern
+ *   - `Array`: File path must match at least one condition
+ * @param file - The file path to check
+ * @returns `true` if the file matches the condition, `false` otherwise
+ *
+ * @example
+ * ```typescript
+ * // String matching
+ * matchCondition('node_modules', '/path/to/node_modules/pkg') // Returns: true
+ *
+ * // RegExp matching
+ * matchCondition(/\.test\.ts$/, 'file.test.ts') // Returns: true
+ *
+ * // Array matching (OR logic)
+ * matchCondition(['src', /\.tsx$/], 'src/App.tsx') // Returns: true
+ * ```
+ */
 export function matchCondition(condition: Condition, file: string) {
   if (typeof condition === 'string') {
     return file.includes(condition);
@@ -123,6 +178,33 @@ export function matchCondition(condition: Condition, file: string) {
   return false;
 }
 
+/**
+ * Get the mapped file path based on the provided mappings configuration
+ *
+ * This function resolves file paths by applying mapping rules, which is useful for:
+ * - Resolving module aliases (e.g., '@/components' -> 'src/components')
+ * - Mapping node_modules paths to local source paths
+ * - Handling monorepo package paths
+ *
+ * @param file - The original file path to map
+ * @param mappings - Path mapping configuration, can be either:
+ *   - Object: `{ '@/': 'src/', '~': 'node_modules/' }`
+ *   - Array: `[{ find: '@/', replacement: 'src/' }, { find: /^~/, replacement: 'node_modules/' }]`
+ * @returns The mapped file path if a mapping is found and the file exists, otherwise returns the original path
+ *
+ * @example
+ * ```typescript
+ * // Object mapping
+ * getMappingFilePath('@/components/Button.tsx', { '@/': 'src/' })
+ * // Returns: 'src/components/Button.tsx' (if file exists)
+ *
+ * // Array mapping with RegExp
+ * getMappingFilePath('~/lodash/index.js', [
+ *   { find: /^~/, replacement: 'node_modules/' }
+ * ])
+ * // Returns: 'node_modules/lodash/index.js' (if file exists)
+ * ```
+ */
 export function getMappingFilePath(
   file: string,
   mappings?:
@@ -211,6 +293,33 @@ function handlePathWithSlash(path: string) {
   return path.endsWith('/') ? path : `${path}/`;
 }
 
+/**
+ * Check if a file should be excluded from processing
+ *
+ * Determines if a file should be excluded based on exclude/include patterns.
+ * Files in node_modules are always excluded unless explicitly included.
+ *
+ * Logic:
+ * - If file matches exclude pattern AND NOT in include pattern → excluded
+ * - If file matches include pattern → NOT excluded (even if in node_modules)
+ * - node_modules is always in the exclude list by default
+ *
+ * @param file - The file path to check
+ * @param options - Code inspector options containing exclude/include patterns
+ * @returns `true` if the file should be excluded, `false` otherwise
+ *
+ * @example
+ * ```typescript
+ * const options = {
+ *   exclude: [/\.test\.ts$/, 'dist'],
+ *   include: ['src']
+ * };
+ *
+ * isExcludedFile('src/App.test.ts', options) // Returns: true (matches exclude)
+ * isExcludedFile('node_modules/pkg/index.js', options) // Returns: true (node_modules)
+ * isExcludedFile('src/index.ts', options) // Returns: false (in include)
+ * ```
+ */
 export function isExcludedFile(file: string, options: CodeOptions) {
   let exclude = options.exclude || [];
   if (!Array.isArray(exclude)) {
@@ -244,4 +353,54 @@ export function hasWritePermission(filePath: string): boolean {
   } catch (error) {
     return false;
   }
+}
+
+/**
+ * Check if a file should be ignored based on special directives in comments
+ * @param content - The file content to check
+ * @param fileType - The type of file ('vue', 'jsx', 'svelte', or unknown)
+ * @returns true if the file should be ignored, false otherwise
+ */
+export function isIgnoredFile({
+  content,
+  fileType,
+}: {
+  content: string;
+  fileType: 'vue' | 'jsx' | 'svelte' | unknown;
+}): boolean {
+  if (!content) {
+    return false;
+  }
+  const trimmed = content.trimStart();
+  const directives = ['code-inspector-disable', 'code-inspector-ignore'];
+
+  // Vue / Svelte - check HTML comments
+  if (fileType === 'vue' || fileType === 'svelte') {
+    if (trimmed.startsWith('<!--')) {
+      const endIndex = trimmed.indexOf('-->');
+      if (endIndex !== -1) {
+        const body = trimmed.slice(0, endIndex + 3).toLowerCase();
+        return directives.some((d) => body.includes(d));
+      }
+    }
+    return false;
+  }
+
+  // Single line comment (// ...)
+  const lineComment = trimmed.match(/^\/\/\s*([^\n]+)/);
+  if (lineComment) {
+    const body = lineComment[1].toLowerCase();
+    return directives.some((d) => body.includes(d));
+  }
+
+  // Block comment (/* ... */)
+  if (trimmed.startsWith('/*')) {
+    const endIndex = trimmed.indexOf('*/');
+    if (endIndex !== -1) {
+      const body = trimmed.slice(0, endIndex + 2).toLowerCase();
+      return directives.some((d) => body.includes(d));
+    }
+  }
+
+  return false;
 }
